Refactor ember-states/view_states out into ember-viewstates so that states is free of ember-views dependency.
Fix whitespace in ember.json. Curse you vimrc

Fix broken tests and fix stuff mentioned by @wagenet in the pull request.

Migrate StateManager's currentView and view documentation to ember-viewstate instead.
