+++ ember.json
@@ -26,6 +26,7 @@
         "handlebars",
         "ember-views",
         "ember-states",
+        "ember-viewstates",
         "ember-handlebars"
       ]
     }
