+++ packages/ember-viewstates/lib/view_state.js
