+++ packages/ember-viewstates/package.json
@@ -1,7 +1,7 @@
 {
   "name": "ember-states",
-  "summary": "Ember Storyboards",
-  "description": "The Ember state management system",
+  "summary": "Ember Stateful Views",
+  "description": "The Ember state-enhanced views",
   "homepage": "http://emberjs.com",
   "author": "Yehuda Katz",
   "version": "0.9.7.1",
@@ -11,7 +11,10 @@
   },
 
   "dependencies": {
-    "spade":              "~> 1.0"
+    "spade":              "~> 1.0",
+    "jquery":             "~> 1.7",
+    "ember-views": "0.9.7.1",
+    "ember-states": "0.9.7.1"
   },
 
   "dependencies:development": {
@@ -23,7 +26,13 @@
     "bpm_libs.js": {
       "files": ["lib"],
       "modes":        "*"
+    },
+
+    "ember-views/bpm_tests.js": {
+      "files": ["tests"],
+      "modes": ["debug"]
     }
+
   }
 }
